+++ src/test/java/kitchenpos/order/service/OrderServiceTest.java
@@ -1,12 +1,20 @@
 package kitchenpos.order.service;
 
 
+import kitchenpos.common.domain.Price;
 import kitchenpos.common.domain.Quantity;
+import kitchenpos.common.exception.NotFoundException;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProducts;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.*;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTableRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,10 +22,13 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
 
 @DisplayName("주문 서비스 테스트")
@@ -27,7 +38,10 @@ public class OrderServiceTest {
     private OrderService orderService;
 
     @Mock
-    private OrderValidator orderValidator;
+    private OrderTableRepository orderTableRepository;
+
+    @Mock
+    private MenuRepository menuRepository;
 
     @Mock
     private OrderRepository orderRepository;
@@ -37,8 +51,11 @@ public class OrderServiceTest {
     void 주문_생성() {
         final OrderLineItemRequest orderLineItemRequest = new OrderLineItemRequest(1L, 1L);
         final OrderRequest orderRequest = new OrderRequest(1L, null, Collections.singletonList(orderLineItemRequest));
+        final Menu menu = Menu.of("menu", Price.of(BigDecimal.valueOf(5000)), 1L, Collections.singletonList(MenuProduct.of(1L, Quantity.of(1L))));
         final Order order = orderRequest.toOrder();
 
+        given(orderTableRepository.findById(orderRequest.getOrderTableId())).willReturn(Optional.of(OrderTable.of(10, false)));
+        given(menuRepository.findById(orderLineItemRequest.getMenuId())).willReturn(Optional.of(menu));
         given(orderRepository.save(order)).willReturn(order);
 
         OrderResponse response = orderService.create(orderRequest);
@@ -47,6 +64,31 @@ public class OrderServiceTest {
                 () -> assertThat(response.getOrderStatus()).isEqualTo(OrderStatus.COOKING),
                 () -> assertThat(response.getOrderLineItems().size()).isEqualTo(1)
         );
+    }
+
+    @DisplayName("주문 테이블을 미존재 예외")
+    @Test
+    void 주문_테이블_미존재_검증() {
+        OrderRequest orderRequest = new OrderRequest(55L, null, Collections.singletonList(new OrderLineItemRequest(1L, 1L)));
+        given(orderTableRepository.findById(anyLong())).willReturn(Optional.empty());
+
+        Throwable thrown = catchThrowable(() -> orderService.create(orderRequest));
+
+        assertThat(thrown).isInstanceOf(NotFoundException.class)
+                .hasMessage("해당 주문 테이블을 찾을 수 없습니다.");
+    }
+
+    @DisplayName("메뉴 미존재 예외")
+    @Test
+    void 메뉴_미존재_검증() {
+        OrderRequest orderRequest = new OrderRequest(55L, null, Collections.singletonList(new OrderLineItemRequest(1L, 1L)));
+        given(orderTableRepository.findById(orderRequest.getOrderTableId())).willReturn(Optional.of(OrderTable.of(10, false)));
+        given(menuRepository.findById(anyLong())).willReturn(Optional.empty());
+
+        Throwable thrown = catchThrowable(() -> orderService.create(orderRequest));
+
+        assertThat(thrown).isInstanceOf(NotFoundException.class)
+                .hasMessage("해당 메뉴를 찾을 수 없습니다.");
     }
 
     @DisplayName("주문 목록을 조회한다.")
